Added in the ability to rotate over time and to reset all servos
Worked on the ability to reset all servos to 0, this is needed to later on use moveOverTime correctly. It has to know where the servo is positioned, this is only possible by setting them to a known value like 0.

Added the ability to rotate over time, you can now use the method MoveServoOverTime to move one or more servos to a specific state in a specific time. This is usefull when you don't want these really fast sharp movements.