Added in a method for picking up items
Worked on the robot a bit. As a test I wanted it to pick up an item from one position and move it to another position. This was a bit tricky at first but I could just use the delayed servo move methods to move the servos not too quickly so that the robot would not fall over. The PickUpItemRoutine methood quickly moves the arm in to position after which it constantly moves between one position and another, picking up one item and dropping it off at the other position. This method is very specific for this scenario and because the robot doesn't have any sensors it will probably stay this way.