Added in a DataReader
Added in the option to send commands to the arm, like ARM move to move a specific servo to a position in a specific time. The ARM now checks for input and responds with the actions. Therefore I also made the pickuproutine execute only once. I tried to make the servos move a bit smoother with a delay between the commands given from the program, this did not work.

Made a wrapper arround the MoveServosOverTime method called MoveServoOverTime where you can just give one servo and it will automaticly make an array and give that to the MoveServosOverTime method.

Included a SoftReset option to reset all the servos to 0 using the MoveServoOverTime method so that things don't break (by moving too fast).