Added in the ability to move multiple servos to different positions
Changed the MoveServosOverTime method to be able to move multiple servos to different states, you used to be able to move multiple servos to the same position, now you are able to move them to different states, giving more possibilities. This is usefull when for example the PickupItemRoutine is running, the routine devides some movements in to multiple steps to make sure the ARM does not fall over. These steps can now be executed at the same time and therefore I can make sure that these steps do not unbalance the arm.